refactor: tidy the slug-cap change after review
Single-source the composed repo-name length through
contract.AssignmentRepoPrefix, drop the migration re-export double-home so
short-name symbols have one import home, add an isValidShortName predicate
shared by both create-form validators, and trim redundant comments.
